Fix issue 1309
