+++ src/extensions/column-templates/defaults/index.js
@@ -41,7 +41,7 @@ import {
 	oneOneOneOneOneOne,
 	twoTwoTwoTwoTwoTwo,
 	twoTwoTwoTwo,
-} from '../../icons';
+} from '../../../icons';
 
 /**
  * Helpers
@@ -175,7 +175,7 @@ const generateDefaultColumns = (columns, isResponsive = false) => {
 };
 
 // Array of all templates
-const templates = {
+const columnTemplates = {
 	oneColumn: {
 		default: [
 			{
@@ -700,124 +700,4 @@ const templates = {
 	},
 };
 
-/**
- * Get Template Object
- *
- * @param {string} templateName Name of template
- * @return {Object} template Object
- */
-function getTemplateObject(templateName) {
-	const allTemplates = [];
-
-	Object.values(templates).forEach(value => {
-		value.default.forEach(template => {
-			allTemplates.push(template);
-		});
-
-		value.responsive.forEach(template => {
-			allTemplates.push(template);
-		});
-	});
-
-	const template = allTemplates.find(
-		template => template.name === templateName
-	);
-
-	return template;
-}
-
-/**
- * Get columns Number based on the template name
- *
- * @param {string} templateName name of the template
- * @return {Integer} Number of Columns for the corresponding template
- */
-function getNumCol(templateName) {
-	let newTemplateName = templateName;
-
-	if (templateName.indexOf('custom-') !== -1) {
-		newTemplateName = templateName.slice(7);
-	}
-
-	const template = getTemplateObject(newTemplateName);
-
-	return template.sizes.length;
-}
-
-/**
- * get templates based on the number of columns and device
- *
- * @param {Integer} columnsNumber Number of columns
- * @return {Array} Array of templates for the corresponding columns number
- */
-function getTemplates(deviceType = 'general', columnsNumber = undefined) {
-	const responsiveSnappingScreens = ['m', 's', 'xs'];
-
-	const {
-		oneColumn,
-		twoColumns,
-		threeColumns,
-		fourColumns,
-		fiveColumns,
-		sixColumns,
-		sevenColumns,
-		eightColumns,
-	} = templates;
-
-	switch (columnsNumber) {
-		case 1:
-			return oneColumn.default.concat(oneColumn.responsive);
-
-		case 2:
-			if (responsiveSnappingScreens.includes(deviceType)) {
-				return twoColumns.default.concat(twoColumns.responsive);
-			}
-			return templates.twoColumns.default;
-
-		case 3:
-			if (responsiveSnappingScreens.includes(deviceType)) {
-				return threeColumns.default.concat(threeColumns.responsive);
-			}
-			return templates.threeColumns.default;
-
-		case 4:
-			if (responsiveSnappingScreens.includes(deviceType)) {
-				return fourColumns.default.concat(fourColumns.responsive);
-			}
-			return fourColumns.default;
-
-		case 5:
-			if (responsiveSnappingScreens.includes(deviceType)) {
-				return fiveColumns.default.concat(fiveColumns.responsive);
-			}
-			return fiveColumns.default;
-		case 6:
-			if (responsiveSnappingScreens.includes(deviceType)) {
-				return sixColumns.default.concat(sixColumns.responsive);
-			}
-			return sixColumns.default;
-		case 7:
-			if (responsiveSnappingScreens.includes(deviceType)) {
-				return sevenColumns.default.concat(sevenColumns.responsive);
-			}
-			return sevenColumns.default;
-		case 8:
-			if (responsiveSnappingScreens.includes(deviceType)) {
-				return eightColumns.default.concat(eightColumns.responsive);
-			}
-			return eightColumns.default;
-
-		default:
-			return oneColumn.default.concat(
-				twoColumns.default,
-				threeColumns.default,
-				fourColumns.default,
-				fiveColumns.default,
-				sixColumns.default,
-				sevenColumns.default,
-				eightColumns.default
-			);
-	}
-}
-
-export { getTemplates, getTemplateObject, getNumCol };
+export default columnTemplates;
