+++ src/extensions/column-templates/index.js
@@ -1,0 +1,6 @@
+import './store';
+
+export { default as getNumCol } from './getNumCol';
+export { default as getTemplates } from './getTemplates';
+export { default as getTemplateObject } from './getTemplateObject';
+export { default as loadColumnsTemplate } from './loadColumnsTemplate';
