+++ src/extensions/column-templates/store/index.js
@@ -1,0 +1,13 @@
+/**
+ * WordPress dependencies
+ */
+import { registerStore } from '@wordpress/data';
+
+/**
+ * Internal dependencies
+ */
+import reducer from './reducer';
+import actions from './actions';
+import * as selectors from './selectors';
+
+registerStore('maxiBlocks/columns', { reducer, actions, selectors });
