+++ src/extensions/column-templates/store/selectors.js
@@ -1,0 +1,16 @@
+/* eslint-disable import/prefer-default-export */
+/**
+ * WordPress dependencies
+ */
+import { select } from '@wordpress/data';
+
+const { getSelectedBlockClientId } = select('core/block-editor');
+
+export const getContentBackup = (state, clientId) => {
+	const selectedClientId = clientId || getSelectedBlockClientId();
+
+	if (state.contentBackup && state.contentBackup[selectedClientId])
+		return state.contentBackup[selectedClientId];
+
+	return {};
+};
