+++ src/extensions/index.js
@@ -1,9 +1,10 @@
 import './attributes';
 import './custom-data';
 import './dom';
-//We are under CGB boilerplate, which doesn't have it, and we are under old Gutenberg version (9.2) and is not adding this store as it experimental. We will destroy it as soon as this package will be available :)
+// We are under CGB boilerplate, which doesn't have it, and we are under old Gutenberg version (9.2) and is not adding this store as it experimental. We will destroy it as soon as this package will be available :)
 import './reusable-blocks';
 import './save';
 import './store';
 import './styles';
 import './text';
+import './column-templates';
